fix:지도 연결 완료
