+++ lib/screen/test_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:hankkitoktok/component/kakao_map_component.dart';
 import 'package:hankkitoktok/screen/0_login_and_set_address/1_address_setting.dart';
 import 'package:hankkitoktok/screen/1_my_page/1_order_histories_screen.dart';
 import 'package:hankkitoktok/screen/2_home/0_home.dart';
@@ -6,6 +7,7 @@ import 'package:hankkitoktok/screen/2_home/1_home_screen.dart';
 import 'package:hankkitoktok/screen/4_pay_choice/pay_test.dart';
 import 'package:hankkitoktok/screen/0_login_and_set_address/0_login_screen.dart';
 import 'package:hankkitoktok/screen/webview_test.dart';
+import '../models/address/address.dart';
 import '0_login_and_set_address/3_view_address_screen.dart';
 import '3_menu_choice/2_order_screen.dart';
 import '4_my_page/0_my_page_home.dart';
@@ -18,6 +20,13 @@ class TestScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+
+    Address1 address = Address1.init(
+      address: '충청북도 청주시 흥덕구 충대로 1',
+      detailAddress: '충북대',
+      latitude: 36.6298968,
+      longitude: 127.4534557,
+    );
 
     return Scaffold(
       appBar: AppBar(
@@ -35,7 +44,8 @@ class TestScreen extends StatelessWidget {
                   _buildButton("pay", context, PayTest(orderId: "1", price: 1000)),
                   _buildButton("map1", context, DeliveryAddressSettingScreen()),
                   _buildButton("map2", context, ViewAddressScreen()),
-                  _buildButton("webview", context, WebviewTest())
+                  _buildButton("webview", context, WebviewTest()),
+                  _buildButton("kakaoMap", context, KakaoMapComponent(address: address)),
                 ]
             )
         )
